✨Feat: SectionView 생성 및 ContentView 수정
